Unified license text and removed incorrect FSF address.
Licensecheck (which is part of the fedora review process) complains
about a number of files using the old and incorrect FSF address in its
license: ./xxx.c: GPL (v2 or later) (with incorrect FSF address)

All files use GPLv2 which I certainly did not change. There were
basically 2 different ways in which the last paragraph of this license
was used. One contained the old address, therefore I decided to use
the other form. This is now used in all license files.

Signed-off-by: Dr. Tilmann Bubeck <t.bubeck@reinform.de>