requestpacket: add range checks to fix buffer overflow
Calling strlen() on a buffer without knowing if there's a null
terminator is easy to crash.  The strlen() function will simply run
across the buffer end, until it happens to find a null byte in the
following (undefined) data section, or until it gets killed by the
kernel.

Another problem is that the method serverError() does not work.  It
works only if nobody has moved the packetPos yet - which is a wrong
assumption to make.