vnsiclient: use strcpy() instead of strncpy()
The use of strncpy() pretends that care was taken to avoid a buffer
overflow, but the nature of strncpy() was misunderstood by the
original code author.  The size parameter is supposed to specify the
buffer size, not the input string size.  By passing the input string
size, the null terminator is omitted, which is only mitigated by the
fact that the replacement string has the same size as the original
string, so the null terminator is already there.

So let's stop pretending, and do what we were really doing all the
time.