vnsiclient: "timeout" is optional in ChannelStream_Open()
Current pvr.vdr.vnsi causes the "Malformed VNSI packet" error to be
triggered because it omits the "timeout" in the ChannelStream_Open
packet.  It's hard to tell what's right and what's wrong, because
there is no VNSI protocol definition.  This worked before (due to lack
of error handling), so let's make the "timeout" optional.