don't retune when a higher priority task is starting
This eliminates the race which occurs when a retune is triggered while
another recording or streaming task with a higher priority is in the
middle of zapping. To detect this, a dummy receiver (one per device) is
attached to the device. When a higher priority task acquires the device,
it calls DetachAllReceivers(), and both receivers, dummy and real, are
detached. In this case, retune shouldn't be triggered. A dummy receiver
is detached together with the last cLiveReceiver.